Heartbeat while downloading the ensemble
Previously, if downloading the ensemble took a long time then tons of
agents could all decide they want to run an ensemble before any of them
increment `started`. This is due to the delay there was between deciding
to run the ensemble and incrementing `started`. Reduce this delay by
incrementing `started` right after picking the ensemble, and continually
heartbeat until either the ensemble is stopped or we increment `ended`